MakeCSV moved to correct folder according to clean architecture. Refactoring used.
