+++ src/test/java/MakeCSVTest.java
@@ -1,7 +1,5 @@
 import EnterpriseBusinessRules.NewCourse;
 import EnterpriseBusinessRules.Session;
-import FrameworksDrivers.MakeCSV;
-import com.sun.source.tree.Tree;
 import org.junit.Test;
 
 import java.io.IOException;
